+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Implement support for SEP-0007 `pay` operation.
 - Add new method `cosmicLink.insertOperation(index, type, params)`.
 
 ## 1.5.5 - 2019-07-26